project_setting/requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.4
Django==4.2.6
django-health-check==3.17.0
django-kombu==0.9.4
djangorestframework==3.14.0
idna==3.4
kombu==5.3.2
mysql-client==0.0.1
mysqlclient==2.2.0
prompt-toolkit==3.0.39
pycparser==2.21
PyMySQL==1.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
rabbitmq==0.2.0
rabbitmq-server==0.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
vine==5.0.0
wcwidth==0.2.8
